refactor: create IExpertsService
