+++ HwProj.AuthService/HwProj.AuthService.API/Services/IAccountService.cs
@@ -13,17 +13,11 @@ namespace HwProj.AuthService.API.Services
         Task<AccountDataDto[]> GetAccountsDataAsync(string[] userIds);
         Task<AccountDataDto> GetAccountDataByEmailAsync(string email);
         Task<Result<TokenCredentials>> RegisterUserAsync(RegisterDataDTO model);
-        Task<Result> RegisterExpertAsync(RegisterExpertViewModel model, string lecturerId);
         Task<Result> EditAccountAsync(string accountId, EditDataDTO model);
-        Task<Result<bool>> GetIsExpertProfileEdited(string expertAccountId);
-        Task<Result> SetExpertProfileIsEdited(string expertAccountId);
         Task<Result<TokenCredentials>> LoginUserAsync(LoginViewModel model);
-        Task<Result> LoginExpertAsync(TokenCredentials tokenCredentials);
         Task<Result<TokenCredentials>> RefreshToken(string userId);
         Task<Result> InviteNewLecturer(string emailOfInvitedUser);
         Task<IList<User>> GetUsersInRole(string role);
-        Task<ExpertDataDTO[]> GetAllExperts();
-        Task<Result> UpdateExpertTags(string lecturerId, UpdateExpertTagsDTO updateExpertTagsDto);
         Task<Result> RequestPasswordRecovery(RequestPasswordRecoveryViewModel model);
         Task<Result> ResetPassword(ResetPasswordViewModel model);
         Task<GithubCredentials> AuthorizeGithub(string code, string userId);
